simplify the volunteer user flow and its form
